requirements.txt
alabaster==0.7.16
Babel==2.15.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.20.1
Flask==3.0.3
idna==3.7
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
Pygments==2.18.0
python-docs-theme==2024.6
PyYAML==6.0.1
requests==2.32.3
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
urllib3==2.2.2
Werkzeug==3.0.3
